Rename package to `redacted'
This commit renames the package from the former name `secret-mode' to
`redacted'.  The rename was necessary since nothing in the `redacted'
buffer was actually secret or save; the files and buffers are still
unencrypted.  Instead of providing a false sense of security, it's
better to come clear and tell users that all they get is some redacted
text, not necessarily something to actually hide secrets.

Why is the package called `redacted' though, and not `redacted-mode'?
Well, I'd like to add a `redacted-region' feature at some point, and it
would be strange to call that function `redact-mode-region'.