Fix version requirements on Emacs
It turns out that the 'symbol pattern only got added to Emacs in
version 25.1. The previous version accidentally announced that Emacs
24.1 would be fine.

Unfortunately, recent automatic tests have shown that this does not
hold. Sorry for the inconvenience of Emacs 24.1 users. Given that
Emacs 25.1 was release in 2016, the amount of users on the old version
should be limited. It's possible to rewrite the `pcase' call to not
use the pattern, but that's left for another day.

The minor version bump in "0.1.1" is intentional, this is a bugfix,
not a breaking change (as the version requirement was previously
broken).